Remove unsed pub types and replace libc with native types
Approved-by: Will Correia <william.correia@cyber.gc.ca>
Approved-by: Todd Mortimer <richard.mortimer@cse-cst.gc.ca>

+++ benches/transactions.rs
@@ -92,7 +92,7 @@ impl Test {
                             self.connp,
                             &tv_start,
                             data.as_ptr() as *const core::ffi::c_void,
-                            data.len() as u64,
+                            data.len(),
                         );
                         if rc == 3 {
                             // HTP_STREAM_ERROR = 3
@@ -101,7 +101,7 @@ impl Test {
 
                         if rc == 5 {
                             // HTP_STREAM_DATA_OTHER = 5
-                            let consumed = htp_connp_req_data_consumed(self.connp) as usize;
+                            let consumed = htp_connp_req_data_consumed(self.connp);
                             let mut remaining = Vec::with_capacity(data.len() - consumed);
                             remaining.extend_from_slice(&data[consumed..]);
                             in_buf = Some(remaining);
@@ -114,7 +114,7 @@ impl Test {
                                 self.connp,
                                 &tv_start,
                                 out_remaining.as_ptr() as *const core::ffi::c_void,
-                                out_remaining.len() as u64,
+                                out_remaining.len(),
                             );
                             out_buf = None;
                             if rc == 3 {
@@ -128,7 +128,7 @@ impl Test {
                             self.connp,
                             &tv_start,
                             data.as_ptr() as *const core::ffi::c_void,
-                            data.len() as u64,
+                            data.len(),
                         );
                         if rc == 3 {
                             // HTP_STREAM_ERROR = 3
@@ -149,7 +149,7 @@ impl Test {
                                 self.connp,
                                 &tv_start,
                                 in_remaining.as_ptr() as *const core::ffi::c_void,
-                                in_remaining.len() as u64,
+                                in_remaining.len(),
                             );
                             in_buf = None;
                             if rc == 3 {
@@ -167,7 +167,7 @@ impl Test {
                     self.connp,
                     &tv_start,
                     out_remaining.as_ptr() as *const core::ffi::c_void,
-                    out_remaining.len() as u64,
+                    out_remaining.len(),
                 );
                 if rc == 3 {
                     // HTP_STREAM_ERROR = 3
